requirements.txt
opencv-python==4.8.1.78
scipy==1.11.4
Pillow==10.1.0
numpy==1.24.4
tensorflow==2.15.0
keras==2.15.0
pandas==2.1.4
matplotlib==3.8.2
scikit-learn==1.3.2
